backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-filter==23.1
djoser==2.1.0
Pillow==9.0.0
pyshorteners
requests==2.26.0
reportlab